feat: Mapa interno dinamico
